backend/requirements.txt
fastapi==0.115.13
uvicorn==0.34.3
requests==2.31.0
beautifulsoup4==4.7.1
Pillow==9.0.0
moviepy==1.0.3
gTTS==2.5.4
numpy==1.26.4
SQLAlchemy==2.0.41